fix(webui): stop mic resample path truncating float samples to -1/0/1
Int16Array.from() always coerces its mapped output back to an integer,
so on browsers that ignore the requested 16kHz AudioContext rate, the
fractional samples computed for resampling were silently truncated to
-1, 0, or 1 before ever reaching resample() — collapsing captured
speech to near-silent noise. Float32Array.from() preserves them.

Verified independently with a standalone Node script reproducing the
exact code path: the buggy line collapses all non-zero samples to
zero, the fix retains their fractional values.

+++ src/jarvis/webui/static/js/mic.js
@@ -103,7 +103,7 @@ export function createMic({ onState, onError } = {}) {
         let buffer = event.data;
         measure(new Int16Array(buffer));
         if (needsResample) {
-          const asFloat = Int16Array.from(new Int16Array(buffer), (v) => v / 0x8000);
+          const asFloat = Float32Array.from(new Int16Array(buffer), (v) => v / 0x8000);
           const resampled = resample(asFloat, ctx.sampleRate, SAMPLE_RATE);
           const pcm = new Int16Array(resampled.length);
           for (let i = 0; i < resampled.length; i++) {
